requirements.txt
# Core Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0

# Core Frontend Dependencies  
streamlit>=1.28.0
httpx>=0.25.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.6.0
pre-commit>=3.5.0
black>=23.9.0